backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-session==0.8.0
bcrypt==4.1.3
openpyxl==3.1.5
resend==2.3.0
python-dotenv==1.0.1
gunicorn==22.0.0
